Fix segmentation fault, thanks to `gdb`
After downgrading to GNU gdb (GDB) 8.0.1, was able to step debug and locate the problem.
I had not reserved enough memory for the variables holding the memory addresses (`heap_begin` and `current_break`).
Instead of the nasm dd, or double word (32-bit), we need quad words (64-bit) for the addresses.